pyproject.toml
[project]
name = "l1-connector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "load-dotenv>=0.1.0",
    "matplotlib>=3.10.1",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.2.3",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.40",
]
